pyproject.toml
[project]
name = "mangify"
version = "0.1.0"
description = "Descubra, leia e viva histórias incríveis em um só app"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.14.1",
    "ulid>=1.1",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.12.12",
    "taskipy>=1.14.1",
    "testcontainers>=4.13.0",
]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
